// chrome/src/models/das-storage.enum.ts
export enum DasStorageEnum {
    ONLINE_CHECK = 'ONLINE_CHECK',
    BATTERY_CHECK = 'BATTERY_CHECK',
    TIME_TO_DISCARD = 'TIME_TO_DISCARD',
    DONT_DISCARD_PINNED = 'DONT_DISCARD_PINNED',
    DONT_DISCARD_FORMS = 'DONT_DISCARD_FORMS',
    DONT_DISCARD_AUDIO = 'DONT_DISCARD_AUDIO',
    IGNORE_CACHE = 'IGNORE_CACHE',
    ADD_TO_CONTEXT_MENU = 'ADD_TO_CONTEXT_MENU',
    DONT_NAG_ME = 'DONT_NAG_ME',
    WHITELIST = 'WHITELIST',
    SYNC_OPTIONS ='SYNC_OPTIONS',
    DISCARD_AT_STARTUP = 'DISCARD_AT_STARTUP'
}